+++ src/test/java/org/folio/inventory/dataimport/cache/MappingMetadataCacheTest.java
@@ -5,7 +5,6 @@ import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import java.util.Optional;
 import java.util.UUID;
 
-import io.vertx.core.json.JsonObject;
 import org.folio.inventory.common.Context;
 import org.folio.inventory.dataimport.handlers.matching.util.EventHandlingUtil;
 import org.folio.MappingMetadataDto;